open/cave mix control via tile survey

+++ src/net/torvald/terrarum/worlddrawer/FeaturesDrawer.kt
@@ -33,12 +33,12 @@ object FeaturesDrawer {
         TileSurvey.submitProposal(
             TileSurvey.SurveyProposal(
                 "basegame.FeaturesDrawer.coldTiles", 72, 48, 2, 2
-            ) { world, x, y -> BlockCodex[world.getTileFromTerrain(x, y)].hasTag("COLD") }
+            ) { world, x, y -> BlockCodex[world.getTileFromTerrain(x, y)].hasTag("COLD").toInt().toFloat() }
         )
         TileSurvey.submitProposal(
             TileSurvey.SurveyProposal(
                 "basegame.FeaturesDrawer.warmTiles", 72, 48, 2, 2
-            ) { world, x, y -> BlockCodex[world.getTileFromTerrain(x, y)].hasTag("WARM") }
+            ) { world, x, y -> BlockCodex[world.getTileFromTerrain(x, y)].hasTag("WARM").toInt().toFloat() }
         )
     }
